+++ src/Symfony/Component/Webhook/Client/RequestParserInterface.php
@@ -30,7 +30,13 @@ interface RequestParserInterface
      */
     public function parse(Request $request, #[\SensitiveParameter] string $secret): ?RemoteEvent;
 
-    public function createSuccessfulResponse(): Response;
+    /**
+     * @param Request|null $request The original request that was received by the webhook controller
+     */
+    public function createSuccessfulResponse(/* ?Request $request = null */): Response;
 
-    public function createRejectedResponse(string $reason): Response;
+    /**
+     * @param Request|null $request The original request that was received by the webhook controller
+     */
+    public function createRejectedResponse(string $reason/* , ?Request $request = null */): Response;
 }