+++ src/Symfony/Component/Webhook/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Add `PayloadSerializerInterface` with implementations to decouple the remote event handling from the Serializer component
+ * Add optional `$request` argument to `RequestParserInterface::createSuccessfulResponse()` and `RequestParserInterface::createRejectedResponse()`
 
 6.4
 ---